Added blank index file for Bus Schedule.
Bus schedule is a personal app I made to interface with the MTA bus time API in order to notify me when I get within two stops of the stop I wish to get off at in the morning on the way to my school. Likely will not remove personalization until much later, if ever. Toggle button for getting on bus, change button for switching to another bus, if bus # is wrong, and a refresh button for manually refreshing the times for the bus arriving, and the next bus.